+++ wechat-user/src/main/java/top/nextnet/service/PurchaseConfirmationGateway.java
@@ -1,5 +1,5 @@
 package top.nextnet.service;
 
 public interface PurchaseConfirmationGateway {
-    void confirmWeChatPurchase(int idWeChatPurchase);
+    void confirmWeChatPurchase(Long idWeChatPurchase);
 }
